criando uma calculadora

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -7,7 +7,7 @@ public class Calculadora {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		
-		int a, b;
+		double a, b;
 		
 		System.out.println("Digite o primeiro valor: ");
 		a = scan.nextInt();
@@ -15,10 +15,10 @@ public class Calculadora {
 		b = scan.nextInt();
 		
 		//atalho; ctrl+1 ou ctrl+2+L
-		int soma = soma(a,b);
-		int subtracao = subtracao(a,b);
-		int multiplicacao = multiplicacao(a,b);
-		int divisao = divisao(a,b);
+		double soma = soma(a,b);
+		double subtracao = subtracao(a,b);
+		double multiplicacao = multiplicacao(a,b);
+		double divisao = divisao(a,b);
 		
 		//atalho; ctrl+alt+ seta baixo
 		System.out.println(soma);
@@ -28,16 +28,16 @@ public class Calculadora {
 		
 	}
 	
-	public static int soma(int a, int b) {
+	public static double soma(double a, double b) {
 		return a + b;
 	}
-	public static int subtracao(int a, int b) {
+	public static double subtracao(double a, double b) {
 		return a - b;
 	}
-	public static int divisao(int a, int b) {
+	public static double divisao(double a, double b) {
 		return a / b;
 	}
-	public static int multiplicacao(int a, int b) {
+	public static double multiplicacao(double a, double b) {
 		return a * b;
 	}
